backend/app/requirement.txt
fastapi
uvicorn
pillow
python-multipart
torch
torchvision
numpy